requirements.txt
asgiref
chardet
colorama
Django
djangorestframework
djangorestframework-simplejwt
drf-nested-routers
drf-yasg
inflection
packaging
Pillow
psycopg2-binary
PyJWT
python-dateutil
pytz
PyYAML
reportlab
qrcode
python-barcode
setuptools
six
sqlparse
tzdata
uritemplate
xmltodict
python-dotenv
celery
redis
